backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
feedparser==6.0.10
pymongo==4.6.0
python-dotenv==1.0.0
schedule==1.2.0
pydantic==2.4.2
httpx==0.25.0
requests==2.31.0 